refactor(test): extract DBAL schema compat logic into trait
