+++ phpstan-6.7.0.0.neon.dist
@@ -101,7 +101,7 @@ parameters:
 
         - message: '#Call to an undefined method Doctrine\\DBAL\\Schema\\(Column|Index)::getObjectName\(\)#'
           identifier: method.notFound
-          path: tests/Migration/CompatMigrationTestCase.php
+          path: tests/Helper/CompatSchemaTrait.php
 
         - message: '#Call to an undefined method Doctrine\\DBAL\\Schema\\AbstractSchemaManager<Doctrine\\DBAL\\Platforms\\AbstractPlatform>::introspectTable(Columns|Indexes)ByUnquotedName\(\)#'
-          path: tests/Migration/CompatMigrationTestCase.php
+          path: tests/Helper/CompatSchemaTrait.php
