+++ tests/Migration/Migration1692001928VaultTokenTest.php
@@ -10,19 +10,22 @@ namespace Swag\PayPal\Test\Migration;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
 use Swag\PayPal\Migration\Migration1692001928VaultToken;
 use Swag\PayPal\Migration\Migration1706111604AddCustomerIdToVault;
+use Swag\PayPal\Test\Helper\CompatSchemaTrait;
 
 /**
  * @internal
  */
 #[Package('checkout')]
 #[CoversClass(Migration1692001928VaultToken::class)]
-class Migration1692001928VaultTokenTest extends CompatMigrationTestCase
+class Migration1692001928VaultTokenTest extends TestCase
 {
+    use CompatSchemaTrait;
     use DatabaseTransactionBehaviour;
     use KernelTestBehaviour;
