+++ tests/Migration/Migration1706111604AddCustomerIdToVaultTest.php
@@ -10,18 +10,21 @@ namespace Swag\PayPal\Test\Migration;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\DatabaseTransactionBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
 use Swag\PayPal\Migration\Migration1706111604AddCustomerIdToVault;
+use Swag\PayPal\Test\Helper\CompatSchemaTrait;
 
 /**
  * @internal
  */
 #[Package('checkout')]
 #[CoversClass(Migration1706111604AddCustomerIdToVault::class)]
-class Migration1706111604AddCustomerIdToVaultTest extends CompatMigrationTestCase
+class Migration1706111604AddCustomerIdToVaultTest extends TestCase
 {
+    use CompatSchemaTrait;
     use DatabaseTransactionBehaviour;
     use KernelTestBehaviour;
